requirements.txt
click==8.1.7
matplotlib==3.9.1
numpy==1.26.4
PennyLane==0.37.0
pytest==8.3.2
scikit-learn==1.5.1
torch==2.4.0
optuna
pyyaml
pandas
